LayoutTests/platform/mac/editing/spelling/autocorrection-blockquote-crash-expected.txt

This test checks that markers are correct when auto correcting in the blockquote. If you type "n" and " ", there should be blue dots under the word: "information". This test also makes sure that the replacement does not cause breaking out of blockquotes. 
Note, this test can fail due to user specific spell checking data. If the user has previously dismissed 'information' as the correct spelling of 'infomation' several times, the spell checker will not provide 'information' as a suggestion anymore. To fix this, remove all files in ~/Library/Spelling. and kill AppleSpell.service

PASS internals.hasAutocorrectedMarker(document, 5, 11) is true
PASS successfullyParsed is true

TEST COMPLETE

would this 
test information 
make a difference?
